Move scss tests into individual directories
