+++ test/integration/scss/scss-fixtures/hmr-module/hmr-module.test.js
@@ -1,0 +1,52 @@
+/* eslint-env jest */
+
+import { remove } from 'fs-extra'
+import { File, findPort, killApp, launchApp, waitFor } from 'next-test-utils'
+import webdriver from 'next-webdriver'
+import { join } from 'path'
+
+describe('Can hot reload CSS Module without losing state', () => {
+  const appDir = __dirname
+
+  let appPort
+  let app
+  beforeAll(async () => {
+    await remove(join(appDir, '.next'))
+    appPort = await findPort()
+    app = await launchApp(appDir, appPort)
+  })
+  afterAll(async () => {
+    await killApp(app)
+  })
+
+  it('should update CSS color without remounting <input>', async () => {
+    const browser = await webdriver(appPort, '/')
+
+    const desiredText = 'hello world'
+    await browser.elementById('text-input').type(desiredText)
+    expect(await browser.elementById('text-input').getValue()).toBe(desiredText)
+
+    const currentColor = await browser.eval(
+      `window.getComputedStyle(document.querySelector('#verify-red')).color`
+    )
+    expect(currentColor).toMatchInlineSnapshot(`"rgb(255, 0, 0)"`)
+
+    const cssFile = new File(join(appDir, 'pages/index.module.scss'))
+    try {
+      cssFile.replace('$var: red', '$var: purple')
+      await waitFor(2000) // wait for HMR
+
+      const refreshedColor = await browser.eval(
+        `window.getComputedStyle(document.querySelector('#verify-red')).color`
+      )
+      expect(refreshedColor).toMatchInlineSnapshot(`"rgb(128, 0, 128)"`)
+
+      // ensure text remained
+      expect(await browser.elementById('text-input').getValue()).toBe(
+        desiredText
+      )
+    } finally {
+      cssFile.restore()
+    }
+  })
+})
