+++ test/integration/scss/scss-fixtures/unused/unused.test.js
@@ -1,48 +1,12 @@
 /* eslint-env jest */
 
 import { remove } from 'fs-extra'
-import { File, findPort, killApp, launchApp, waitFor } from 'next-test-utils'
+import { findPort, killApp, launchApp, waitFor, File } from 'next-test-utils'
 import webdriver from 'next-webdriver'
 import { join } from 'path'
 
-const fixturesDir = join(__dirname, '..', 'scss-fixtures')
-
-describe('Has CSS in computed styles in Development', () => {
-  const appDir = join(fixturesDir, 'multi-page')
-
-  beforeAll(async () => {
-    await remove(join(appDir, '.next'))
-  })
-
-  let appPort
-  let app
-  beforeAll(async () => {
-    appPort = await findPort()
-    app = await launchApp(appDir, appPort)
-  })
-  afterAll(async () => {
-    await killApp(app)
-  })
-
-  it('should have CSS for page', async () => {
-    let browser
-    try {
-      browser = await webdriver(appPort, '/page2')
-
-      const currentColor = await browser.eval(
-        `window.getComputedStyle(document.querySelector('.blue-text')).color`
-      )
-      expect(currentColor).toMatchInlineSnapshot(`"rgb(0, 0, 255)"`)
-    } finally {
-      if (browser) {
-        await browser.close()
-      }
-    }
-  })
-})
-
 describe('Body is not hidden when unused in Development', () => {
-  const appDir = join(fixturesDir, 'unused')
+  const appDir = __dirname
 
   beforeAll(async () => {
     await remove(join(appDir, '.next'))
@@ -75,7 +39,7 @@ describe('Body is not hidden when unused in Development', () => {
 })
 
 describe('Body is not hidden when broken in Development', () => {
-  const appDir = join(fixturesDir, 'unused')
+  const appDir = __dirname
 
   let appPort
   let app
